move functions to utils/data.py
