+++ src/gluonnlp/utils/parameter.py
@@ -24,11 +24,11 @@ __all__ = ['clip_grad_global_norm', 'save_parameters',
 import warnings
 import os
 import tempfile
-import logging
 
 import numpy as np
 from mxnet import nd
 from .. import _constants as C
+from .data import _TempFilePath, _transfer_file_s3
 
 def clip_grad_global_norm(parameters, max_norm, check_isfinite=True):
     """Rescales gradients of parameters so that the sum of their 2-norm is smaller than `max_norm`.
@@ -108,7 +108,7 @@ def _s3_compatible_save_load(is_save, save_load_method, filename, *args, **kwarg
     """Dispatch function for save load with s3."""
     if C.S3_PREFIX in filename:
         # create temp dir
-        with TempFilePath() as temp_path:
+        with _TempFilePath() as temp_path:
             if is_save:
                 # save model
                 save_load_method(temp_path, *args, **kwargs)
@@ -160,30 +160,6 @@ def save_parameters(model, filename):
     """
     _s3_compatible_save_load(True, model.save_parameters, filename)
 
-def _transfer_file_s3(filename, s3_filename, upload=True):
-    """Transfer a file between S3 and local file system."""
-    try:
-        import boto3
-    except ImportError:
-        raise ImportError('boto3 is required to support s3 URI. Please install'
-                          'boto3 via `pip install boto3`')
-    # parse s3 uri
-    prefix_len = len(C.S3_PREFIX)
-    bucket_idx = s3_filename[prefix_len:].index('/') + prefix_len
-    bucket_name = s3_filename[prefix_len:bucket_idx]
-
-    # filename after the bucket, excluding '/'
-    key_name = s3_filename[bucket_idx + 1:]
-
-    log_level = logging.getLogger().getEffectiveLevel()
-    logging.getLogger().setLevel(logging.INFO)
-    # upload to s3
-    s3 = boto3.client('s3')
-    if upload:
-        s3.upload_file(filename, bucket_name, key_name)
-    else:
-        s3.download_file(bucket_name, key_name, filename)
-    logging.getLogger().setLevel(log_level)
 
 def load_states(trainer, fname):
     """Loads trainer states (e.g. optimizer, momentum) from a file.
@@ -225,19 +201,3 @@ def save_states(trainer, fname):
     `lr_mult` and `wd_mult`) will not be saved.
     """
     _s3_compatible_save_load(True, trainer.save_states, fname)
-
-class TempFilePath(object):
-    """A TempFilePath that provides a path to a temporarily file, and automatically
-    cleans up the temp file at exit.
-    """
-    def __init__(self):
-        self.temp_dir = os.path.join(tempfile.gettempdir(), str(hash(os.times())))
-        if not os.path.exists(self.temp_dir):
-            os.makedirs(self.temp_dir)
-
-    def __enter__(self):
-        self.temp_path = os.path.join(self.temp_dir, str(hash(os.times())))
-        return self.temp_path
-
-    def __exit__(self, exec_type, exec_value, traceback):
-        os.remove(self.temp_path)
